pyproject.toml
[tool.poetry]
name = "meal-planner"
version = "0.1.0"
description = ""
authors = ["Gwendolyn Faraday <gwenfaraday@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-aiplatform = "^1.35.0"
colorama = "^0.4.6"
pyfiglet = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
